Refactored BitmapSourceConverter/WriteableBitmapConverter. Changed visibility.
The ToMat conversions now work directly with the base class:
BitmapSource. This allows BitmapSource and WriteableBitmap to be
converted to the OpenCV Mat.

The ToBitmapSource methods now use the Conversions from within
WriteableBitmap.

Also changed visibility of WriteableBitmapConverter to allow usage of
the extension methods defined there.

Remarks:
- Did not refactor the Pixelformat mapping.
- Did not alter the actual conversion code.
- The removal of the ToMat methods from WriteableBitmapConverter would
be a breaking change if there wasn't that visibility bug that prevented
the use of any extension methods defined in that class.